Add files via upload
Python program for plotting the bifurcation diagram resulting from the logistic equation. First defines a function for the logistic equation and then implements this in a for loop for a range of values of the independent variable (r). The bifurcation diagram is a very important tool for understanding chaotic dynamics. It also exhibits fractal properties.